% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCSB_API.R
\name{download_structures}
\alias{download_structures}
\title{Download mmCIF or PDB files}
\usage{
download_structures(
  pdb_ids,
  format = "cif",
  dest_dir = "results/structures",
  verbose = TRUE
)
}
\arguments{
\item{pdb_ids}{Vector of PDB IDs}

\item{format}{Either "cif" or "pdb"}

\item{dest_dir}{Directory to store downloaded files}

\item{verbose}{Show progress}
}
\description{
Download mmCIF or PDB files
}
