% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_pdb.R
\name{fetch_pdb}
\alias{fetch_pdb}
\title{Fetch PDB Structure from RCSB PDB}
\usage{
fetch_pdb(pdb_id)
}
\arguments{
\item{pdb_id}{A string specifying the PDB ID (e.g., "1a2b").}
}
\value{
File path to the downloaded PDB file.
}
\description{
Downloads PDB files from the RCSB database.
}
