% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCSB_API.R
\name{read_query}
\alias{read_query}
\title{Read JSON query from file}
\usage{
read_query(filepath)
}
\arguments{
\item{filepath}{Path to the JSON query file}
}
\value{
Parsed list object of the query
}
\description{
Read JSON query from file
}
