% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCSB_API.R
\name{save_results}
\alias{save_results}
\title{Save metadata and results}
\usage{
save_results(result, output_dir = "results", query_file = NULL)
}
\arguments{
\item{result}{Parsed API result}

\item{output_dir}{Output directory}

\item{query_file}{Path to original query file}
}
\description{
Save metadata and results
}
