% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCSB_API.R
\name{submit_query}
\alias{submit_query}
\title{Submit query to RCSB with retry logic}
\usage{
submit_query(query, retries = 5)
}
\arguments{
\item{query}{List query object}

\item{retries}{Max retry attempts on failure}
}
\value{
Parsed result list
}
\description{
Submit query to RCSB with retry logic
}
