% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_pdb.R
\name{visualize_pdb}
\alias{visualize_pdb}
\title{Visualize PDB Structure in 3D}
\usage{
visualize_pdb(pdb_path)
}
\arguments{
\item{pdb_path}{Path to the PDB file.}
}
\description{
Displays a 3D interactive plot of the PDB structure.
}
